Fix fieldMetadata sync validation exceptions caught in exception handler
