+++ packages/twenty-server/src/engine/metadata-modules/relation-metadata/utils/relation-metadata-graphql-api-exception-handler.util.ts
@@ -8,8 +8,13 @@ import {
   RelationMetadataException,
   RelationMetadataExceptionCode,
 } from 'src/engine/metadata-modules/relation-metadata/relation-metadata.exception';
+import { InvalidMetadataException } from 'src/engine/metadata-modules/utils/exceptions/invalid-metadata.exception';
 
 export const relationMetadataGraphqlApiExceptionHandler = (error: Error) => {
+  if (error instanceof InvalidMetadataException) {
+    throw new UserInputError(error.message);
+  }
+
   if (error instanceof RelationMetadataException) {
     switch (error.code) {
       case RelationMetadataExceptionCode.RELATION_METADATA_NOT_FOUND:
