+++ packages/twenty-server/src/engine/metadata-modules/utils/exceptions/invalid-metadata.exception.ts
@@ -1,0 +1,5 @@
+export class InvalidMetadataException extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
